4.4.2 ZU GPE proprietary files

+++ rootdir/system/etc/init.qcom.wifi.sh
@@ -1,6 +1,6 @@
 #!/system/bin/sh
 # Copyright (c) 2010-2013, The Linux Foundation. All rights reserved.
-# Copyright (C) 2013 Sony Mobile Communications AB.
+# Copyright (c) 2013 Sony Mobile Communications AB.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions are
@@ -61,7 +61,7 @@ trigger_wcnss()
             # caldata is available at userspace.
             if [ -e /data/misc/wifi/WCNSS_qcom_wlan_cal.bin ]; then
                 calparm=`ls /sys/module/wcnsscore/parameters/has_calibrated_data`
-                if [ -e $calparm ]; then
+                if [ -e $calparm ] && [ ! -e /data/misc/wifi/WCN_FACTORY ]; then
                     echo 1 > $calparm
                 fi
             fi
@@ -98,8 +98,6 @@ case "$target" in
     if [ -e /sys/bus/platform/drivers/msm_hsic_host ]; then
        if [ ! -L /sys/bus/usb/devices/1-1 ]; then
            echo msm_hsic_host > /sys/bus/platform/drivers/msm_hsic_host/unbind
-       else
-           echo auto > /sys/bus/usb/devices/1-1/power/control
        fi
 
        chown system.system /sys/bus/platform/drivers/msm_hsic_host/bind
@@ -132,7 +130,7 @@ case "$target" in
                 for device in $sdio_devices; do
                     if [ $ven_idx -eq $dev_idx ]; then
                         case "$device" in
-                        "0x0400" | "0x0401")
+                        "0x0400" | "0x0401" | "0x0402")
                             wlanchip="AR6004-SDIO"
                             ;;
                         *)
@@ -190,6 +188,7 @@ case "$target" in
       echo msm_hsic_host > /sys/bus/platform/drivers/msm_hsic_host/unbind
       setprop wlan.driver.ath 2
       setprop qcom.bluetooth.soc ath3k
+      btsoc="ath3k"
       rm  /system/lib/modules/wlan.ko
       ln -s /system/lib/modules/ath6kl-3.5/ath6kl_usb.ko \
 		/system/lib/modules/wlan.ko
@@ -207,6 +206,7 @@ case "$target" in
       "AR6004-SDIO")
       setprop wlan.driver.ath 2
       setprop qcom.bluetooth.soc ath3k
+      btsoc="ath3k"
       # Chown polling nodes as needed from UI running on system server
       chmod 0200 /sys/devices/msm_sdcc.1/polling
       chmod 0200 /sys/devices/msm_sdcc.2/polling
@@ -227,6 +227,7 @@ case "$target" in
 		/system/etc/firmware/ath6k/AR6004/hw1.3/fw.ram.bin
       ln -s /system/etc/firmware/ath6k/AR6004/hw1.3/bdata.bin_sdio \
 		/system/etc/firmware/ath6k/AR6004/hw1.3/bdata.bin
+      rm /system/etc/firmware/ath6k/AR6004/hw3.0/bdata.bin
       ln -s /system/etc/firmware/ath6k/AR6004/hw3.0/bdata.bin_sdio \
                 /system/etc/firmware/ath6k/AR6004/hw3.0/bdata.bin
       ;;
@@ -239,15 +240,16 @@ case "$target" in
       ln -s /system/lib/modules/pronto/pronto_wlan.ko \
 		/system/lib/modules/wlan.ko
       # Populate the writable driver configuration file
-      #if [ ! -e /data/misc/wifi/WCNSS_qcom_cfg.ini ]; then
-      #    cp /system/etc/wifi/WCNSS_qcom_cfg.ini \
-      #		/data/misc/wifi/WCNSS_qcom_cfg.ini
-      #    chmod 660 /data/misc/wifi/WCNSS_qcom_cfg.ini
-      #fi
+#      if [ ! -e /data/misc/wifi/WCNSS_qcom_cfg.ini ]; then
+#          cp /system/etc/wifi/WCNSS_qcom_cfg.ini \
+#		/data/misc/wifi/WCNSS_qcom_cfg.ini
+#          chown system:wifi /data/misc/wifi/WCNSS_qcom_cfg.ini
+#          chmod 660 /data/misc/wifi/WCNSS_qcom_cfg.ini
+#      fi
 
       # The property below is used in Qcom SDK for softap to determine
       # the wifi driver config file
-      #setprop wlan.driver.config /data/misc/wifi/WCNSS_qcom_cfg.ini
+#      setprop wlan.driver.config /data/misc/wifi/WCNSS_qcom_cfg.ini
       setprop wlan.driver.config /system/etc/firmware/wlan/prima/WCNSS_qcom_cfg.ini
 
       # Trigger WCNSS platform driver
@@ -454,9 +456,6 @@ case "$target" in
     *)
       ;;
 esac
-
-# Load cfg80211 kernel module
-insmod /system/lib/modules/cfg80211.ko
 
 # Run audio init script
 /system/bin/sh /system/etc/init.qcom.audio.sh "$target" "$btsoc"
